se completa modulo de finanzas  y subida de archivos

+++ controllers/FinanceController.js
@@ -1,4 +1,6 @@
 import fs from 'fs'
+import { join, extname } from 'path'
+import { cwd } from 'process'
 
 export default class FinanceController {
     
@@ -68,6 +70,29 @@ export default class FinanceController {
 
     async delete(req,res){
         const id = req.params['id']
+        if('costumer_id' in req.query == false){
+            return res.json({
+                ok:false,
+                message:"Debe enviar el ID del cliente",
+            })
+        }
+        if('filename' in req.query == false){
+            return res.json({
+                ok:false,
+                message:"Debe enviar el nombre del archivo",
+            })
+        }
+
+        const filePath = join(cwd(),'storage','finances', req.query['costumer_id'], req.query['filename'])
+
+        if(!fs.existsSync(filePath)){
+            return res.status(404).json({
+                ok:false,
+                message:'El archivo no existe'
+            })
+        }
+        fs.unlinkSync(filePath)
+
         try {
             await this.financeService.delete({id})
             return res.json({
@@ -147,6 +172,43 @@ export default class FinanceController {
                 error
             })
         }
+    }
+
+    async findFinanceFile(req,res){
+        const { filename } = req.params;
+        const { costumer_id, key } = req.query;
+
+        if(!key){
+            return res.status(404).json({
+                ok:false,
+                message:'Debe enviar la API KEY'
+            })
+        }
+
+        if(key != `${process.env.SECRET_KEY}|${process.env.API_TOKEN}`){
+            return res.status(401).json({
+                ok:false,
+                message:'No tiene permiso para solicitar este archivo'
+            })
+        }
+
+        if(!costumer_id){
+            return res.status(404).json({
+                ok:false,
+                message:'Debe enviar el ID del cliente'
+            })
+        }
+
+        const filePath = join(cwd(),'storage','finances', costumer_id, filename)
+
+        if(!fs.existsSync(filePath)){
+            return res.status(404).json({
+                ok:false,
+                message:'El archivo no existe'
+            })
+        }
+
+        res.sendFile(filePath)
     }
 
 };
